ai/requirements.txt
fastapi
uvicorn[standard]
ultralytics
pillow
requests
python-multipart
deepface
opencv-python
tf-keras


